Refactor LLM Task
